# specs/001-multi-tenant-user-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Multi-Tenant User Service API
  description: >
    A production-grade, multi-tenant User Service API integrating with Supabase. 
    Provides full lifecycle management for tenants, users, roles, and permissions 
    enforcing strict tenant isolation via JWT tokens.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Current API version
...
paths:
...
  /tenants:
    post:
      tags: [Tenants]
      summary: Create a new tenant (System Admin only)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TenantCreate'
      responses:
        '201':
          description: Tenant created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /tenants/{id}:
    get:
      tags: [Tenants]
      summary: Retrieve a specific tenant's details
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PathId'
      responses:
        '200':
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PathId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: The unique UUID of the resource
  schemas:
...
    Tenant:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, suspended]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TenantCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [active, suspended]
          default: active
...
    ErrorModel:
      type: object
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "Invalid input parameters"
        details:
          type: object
          additionalProperties: true
        trace_id:
          type: string
          format: uuid
  responses:
...
    ForbiddenError:
      description: Insufficient permissions for this action (RBAC / Tenant Isolation)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'
    NotFoundError:
      description: Resource not found or isolated in another tenant
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorModel'